feat(view-collections): add support for view collections
View collections are read-only collections that differ from the default
base collection in PocketBase. View collections don't necessarily have a
`created` or `updated` column. Until now the loader needed at least the
`updated` column, to enable incremental updates.

With this commit, view collections without the `updated` column are now
also supported. The loader will check for the column to be present
during the type check and while loading the entries. This ensures that
the generated types are correct again and that the request to load the
entries doesn't fail because of a missing column. This also means that
**incremental updates are not possible** with view collections that
don't have the `updated` column.

_Side note: `updated` is always a date or `undefined`. Renaming another
column to `updated` is only possible if that column is of type `date`.