feat(files): transform file names to file urls
This commit adds built in support for files in PocketBase collections.
The loader now automatically transforms file names retured by the
PocketBase API to URLs where these files can be accessed.

While testing this change, I found another bug in the incremental
updates, where unchanged entries would not update the file URL with the
new values, due to the same digest value. This is why the loader now
forces a complete refresh when the version of this package is updated.