feat(json-fields): add better support for json fields
Before this commit, `json` fields would only return empty objects.
This commit now changes this to return the full object of type
`unknown`. While this is not 100% correct, we don't have any other
option out of the box.

If a user also wants their `json` fields to be typed in the collection,
they can provide a record of schemas to the new `jsonSchemas` option.
These schemas will then be used to generate types for the given field.
While this makes the type generation more correct it also makes the type
checking stricter for all entries.