feat(*): add support for PocketBase 0.23.0
BREAKING CHANGE: This also removes support for PocketBase 0.22.
There are a lot of breaking changes in this new version of PocketBase,
e.g. new endpoint for login, new collection schema format, etc.

Since this version already brings a lot of changes, I used this chance
to refactor some of the internals and configuration options. Please
refer to the new README for more details.