feat(refresh): check refresh context for mentioned collection
When a collection to refresh was mentioned and the source that triggered
the reload was the `astro-integration-pocketbase`, we now check if the
collection name matches the one registered in the loader. This allows us
to skip the refresh if the collection name is not the same, thus
preventing unneccesary reloading operations.

Closes #23